+++ codershq/templates/users/admin/change_list.html
@@ -1,0 +1,7 @@
+{% extends "admin/change_list.html" %}
+
+    {{ block.super }}
+
+{% block object-tools %}
+    <button>Import</button>
+{% endblock %}
