+++ codershq/users/admin.py
@@ -4,6 +4,8 @@ from django.contrib.auth import get_user_model
 from django.utils.translation import gettext_lazy as _
 from django.http import HttpResponse
 from codershq.users.forms import UserChangeForm, UserCreationForm
+import csv
+
 User = get_user_model()
 
 
@@ -33,3 +35,29 @@ class UserAdmin(auth_admin.UserAdmin):
     )
     list_display = ["username", "name", "is_superuser"]
     search_fields = ["name"]
+    
+    # change_list_template = '/templates/users/admin/change_list.html'
+    actions = ['export', ]
+
+    def export(self,request,queryset):
+        meta = self.model._meta
+
+        # field names to be exported
+        # use [field.name for field in meta.fields] to import all the data of specific user
+        # here it only uploads the name and email 
+        fieldnames = ["username","email"]
+
+        response = HttpResponse(content_type='text/csv')
+
+        response['Content-Disposition'] = 'attachment; filename="UsersData.csv"'
+
+        writer = csv.writer(response)
+
+        writer.writerow(fieldnames) #heading in csv files
+
+        for obj in queryset:
+            row = writer.writerow([getattr(obj, field) for field in fieldnames])
+
+        return response
+
+    export.short_description ="Export to csv"
